+++ extensions/tiktok/src/main/java/app/morphe/extension/tiktok/settings/preference/categories/CommentsPreferenceCategory.java
@@ -20,7 +20,8 @@ public class CommentsPreferenceCategory extends ConditionalPreferenceCategory {
         return SettingsStatus.commentToolsEnabled
                 || SettingsStatus.commentTranslationEnabled
                 || SettingsStatus.hideCommentQuickReactionsEnabled
-                || SettingsStatus.copyCommentsWithoutUsernameEnabled;
+                || SettingsStatus.copyCommentsWithoutUsernameEnabled
+                || SettingsStatus.hideCommentEggsEnabled;
     }
 
     @Override